working BB API
need to check dashboard for orders?

next step: getting bb id and password from backend
